Reorganised alamac code to be nested under certain abstractions
