+++ src/packages/almanac/src/configuration/index.ts
@@ -1,0 +1,1 @@
+export * from './options';
