+++ src/packages/almanac/src/otel.ts
@@ -1,2 +1,1 @@
-export { otelTracer } from './strategies/otelTracer';
-export { otelMeter } from './strategies/otelMeter';
+export * from './providers/otel';
