+++ src/packages/almanac/src/scopes/index.ts
@@ -1,0 +1,1 @@
+export * from './scope';
