+++ src/packages/almanac/src/testing/index.ts
@@ -1,0 +1,6 @@
+export * from './collectedSpan';
+export * from './collectingMeter';
+export * from './collectingTracer';
+export * from './spanAssertions';
+export * from './testing';
+
