--- src/packages/almanac/src/types.ts
@@ -1,3 +1,0 @@
-export type TracePropagation = 'links' | 'propagate';
-
-export type AttributeTarget = 'mainSpan' | 'currentSpan' | 'both';
